RESTRUTURAÇÃO DE COMPONENTES E ROTAS
