+++ src/context/auth-context.js
@@ -1,6 +1,5 @@
 import { createContext, useState, useContext } from "react";
 import { useLocation, Navigate } from "react-router-dom";
-
 import { setInStorage, login } from "../services/auth";
 
 const AuthContext = createContext(null);
