--- src/routes/Feed.js
@@ -1,50 +1,0 @@
-import { Textarea, Box, Avatar } from "@chakra-ui/react";
-
-function Feed() {
-  return (
-    <Box color="black" w={{ base: "full", md: 664 }}>
-      <Textarea
-        placeholder="O que está acontecendo?"
-        Avatar="https://bit.ly/dan-abramov"
-      >
-        <Avatar
-          size="2xs"
-          name="Dan Abrahmov"
-          src="https://bit.ly/dan-abramov"
-        />
-      </Textarea>
-      There are many benefits to a joint design and development system. Not only
-      does it bring benefits to the design team, but it also brings benefits to
-      engineering teams. It makes sure that our experiences have a consistent
-      look and feel, not just in our design specs, but in
-      production.ffffffffffffffff There are many benefits to a joint design and
-      development system. Not only does it bring benefits to the design team,
-      but it also brings benefits to engineering teams. It makes sure that our
-      experiences have a consistent look and feel, not just in our design specs,
-      but in production.ffffffffffffffff There are many benefits to a joint
-      design and development system. Not only does it bring benefits to the
-      design team, but it also brings benefits to engineering teams. It makes
-      sure that our experiences have a consistent look and feel, not just in our
-      design specs, but in production.ffffffffffffffff There are many benefits
-      to a joint design and development system. Not only does it bring benefits
-      to the design team, but it also brings benefits to engineering teams. It
-      makes sure that our experiences have a consistent look and feel, not just
-      in our design specs, but in production.ffffffffffffffff There are many
-      benefits to a joint design and development system. Not only does it bring
-      benefits to the design team, but it also brings benefits to engineering
-      teams. It makes sure that our experiences have a consistent look and feel,
-      not just in our design specs, but in production.ffffffffffffffff There are
-      many benefits to a joint design and development system. Not only does it
-      bring benefits to the design team, but it also brings benefits to
-      engineering teams. It makes sure that our experiences have a consistent
-      look and feel, not just in our design specs, but in
-      production.ffffffffffffffff There are many benefits to a joint design and
-      development system. Not only does it bring benefits to the design team,
-      but it also brings benefits to engineering teams. It makes sure that our
-      experiences have a consistent look and feel, not just in our design specs,
-      but in production.ffffffffffffffff
-    </Box>
-  );
-}
-
-export default Feed;
