--- src/routes/ProtectedPage.js
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-function ProtectedPage() {
-  return <Title>Protected</Title>;
-}
-
-const Title = styled.h3`
-  color: blue;
-`;
-
-export default ProtectedPage;
